Implemented "Please Select..." on FormComboBox  and split converters into multiple files
