+++ src/ServiceControl.Config/ServiceControl.Config.csproj
@@ -260,6 +260,9 @@
     <Compile Include="Xaml\Controls\CleanWindow.cs" />
     <Compile Include="Xaml\Controls\FormComboBox.cs" />
     <Compile Include="Xaml\Controls\FormTextBox.cs" />
+    <Compile Include="Xaml\Converters\NegaterConverter.cs" />
+    <Compile Include="Xaml\Converters\SelectedIndexToVisibilityConverter.cs" />
+    <Compile Include="Xaml\Converters\StringNullOrEmptyToVisibilityConverter.cs" />
     <Compile Include="Xaml\Native\UnsafeNativeMethods.cs" />
     <Compile Include="Xaml\Native\NativeConstants.cs" />
     <Compile Include="Xaml\Native\ShowWindowCommands.cs" />
