+++ src/ServiceControl.Config/Xaml/Converters/SelectedIndexToVisibilityConverter.cs
@@ -1,0 +1,26 @@
+﻿namespace ServiceControl.Config.Xaml.Converters
+{
+    using System;
+    using System.Globalization;
+    using System.Windows;
+    using System.Windows.Data;
+
+    public class SelectedIndexToVisibilityConverter : System.Windows.Markup.MarkupExtension, IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            var i = System.Convert.ToInt32(value);
+            return i > -1 ? Visibility.Hidden : Visibility.Visible;
+
+        }
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            return null;
+        }
+
+        public override object ProvideValue(IServiceProvider serviceProvider)
+        {
+            return this;
+        }
+    }
+}
